Create Lines of Code
This program lets the user inuput a string of data associated with a designated country.  From it, the user is presented with 4 options out of which, if he follows the designated instructions he will be free to select. from this series of options, the user will be able to inquire on certain counttries oor check if a country is within the data provided. The user's activities will then be appraised by the program and it will react with different results.